Begin to style and reorganize the files to have a hierarchy and grouping. To encourage functional cohesive modules.
