--- disney-reservation-app/src/pages/home/AddReservationForm.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const AddReservationForm = () => <div>Reservation Form</div>
-
-export default AddReservationForm;
