+++ disney-reservation-app/src/pages/home/HomePage.module.css
@@ -1,0 +1,3 @@
+.page {
+    display: flex;
+}
