--- disney-reservation-app/src/pages/home/ReservationEventListView.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const ReservationEventListView = () => {
-
-    return <div>This will be a listview</div>
-};
-
-export default ReservationEventListView;
