+++ disney-reservation-app/src/pages/home/form/AddReservationForm.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import styles from './AddReservationForm.module.css';
+
+const AddReservationForm = () => <div className={styles.form}>Reservation Form</div>
+
+export default AddReservationForm;
