+++ disney-reservation-app/src/pages/home/form/AddReservationForm.module.css
@@ -1,0 +1,4 @@
+.form {
+    min-width: 600px;
+    background-color: green;
+}
