+++ disney-reservation-app/src/pages/home/listview/ReservationEventListView.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import styles from './ReservationEventListView.module.css';
+
+const ReservationEventListView = () => {
+
+    return <div className={styles.listView}>This will be a listview</div>
+};
+
+export default ReservationEventListView;
