+++ disney-reservation-app/src/pages/home/listview/ReservationEventListView.module.css
@@ -1,0 +1,4 @@
+.listView {
+    min-width: 200px;
+    background-color: grey;
+}
